Restructure to focus on holding ICP and tokens
